requirements.txt
dependency-injector>=4.41.0
tomli>=2.0.2
bs4>=0.0.2 
pyodide-py>=0.25.1
aiofiles>=24.1.0
starlette>=0.39.2
supabase>=2.13.0 
python-multipart>=0.0.12
itsdangerous==2.2.0 
websockets>=13.1
paramiko>=3.5.0
Jinja2>=3.1.4
PyJWT>=2.10.1
untangle>=1.2.1
uvicorn>=0.31.0
aiohttp>=3.10.10
ansible-runner>=2.4.1
flet-video==0.1.0
tinycss>=0.4
cerberus>=1.3.7
requests>=2.32.5
aiofiles>=24.1.0
aiohttp>=3.10.10
beautifulsoup4>=0.0.2
cerberus>=1.3.7
itsdangerous==2.2.0
Jinja2>=3.1.4
paramiko>=3.5.0
PyJWT>=2.10.1
python-multipart>=0.0.12
requests>=2.32.5
starlette>=0.39.2
supabase>=2.13.0
untangle>=1.2.1
uvicorn>=0.31.0
websockets>=13.1
psutil>=7.1.0
redis>=7.0.1
lark>=1.3.1
mistql>=0.4.12
networkx>=3.0
htpy>=25.12.0
psycopg2-binary>=2.9.10